Imported from brave-ui: Welcome - make welcome page images exportable
