+++ src/features/welcome/index.ts
@@ -2,16 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Background, BackgroundContainer } from './background'
 import { SkipButton, Bullet } from './button'
-import {
-  BraveImage,
-  PaymentsImage,
-  ImportImage,
-  ShieldsImage,
-  SearchImage,
-  ThemeImage
-} from './image'
 import { Title, Paragraph } from './text'
 import {
   Footer,
@@ -26,16 +17,8 @@ import {
 } from './wrapper'
 
 export {
-  Background,
-  BackgroundContainer,
   SkipButton,
   Bullet,
-  BraveImage,
-  PaymentsImage,
-  ImportImage,
-  ShieldsImage,
-  SearchImage,
-  ThemeImage,
   Title,
   Paragraph,
   Footer,
